+++ script/test.sh
@@ -2,4 +2,4 @@
 set -o errexit -o nounset -o pipefail
 cd "$(dirname "$0")/.."
 
-lein do test, test-cljs
+lein do javac, test, test-cljs